% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ob_timings}
\alias{ob_timings}
\title{Outbreak Timing from Simulations}
\format{
A data frame with 4 columns:
\describe{
\item{R0}{Basic reproductive number (5-10)}
\item{VE}{Vaccine efficacy (0.5-0.9)}
\item{coverage}{Population vaccination coverage (0-1)}
\item{breach_times}{Time to outbreak from simulations (days)}
}
}
\source{
Zenodo database
DOI: \url{https://doi.org/10.5281/zenodo.5802433}
}
\usage{
ob_timings
}
\description{
Raw simulation results containing breach times for each parameter combination.
Each row represents one simulated breach event from computational models
of COVID-19 quarantine systems.
}
\references{
Zachreson, C., Shearer, F. M., Price, D. J., Lydeamore, M. J., McVernon, J.,
McCaw, J., & Geard, N. (2022). COVID-19 in low-tolerance border quarantine
systems: Impact of the Delta variant of SARS-CoV-2.
Science Advances, 8(14), eabm3624.
}
\keyword{datasets}
