% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run COVID-19 Quarantine Shiny App}
\usage{
run_app()
}
\value{
Launches the Shiny application
}
\description{
Launches an interactive Shiny application for exploring COVID-19
quarantine outbreak probability under different vaccination scenarios.
}
\examples{
\dontrun{
  run_app()
}
}
